+++ src/main/java/com/mg105/use_cases/inventory/InventoryInteractor.java
@@ -1,4 +1,4 @@
-package com.mg105.usecase.Inventory;
+package com.mg105.use_cases.inventory;
 
 import com.mg105.entities.BattleCharacter;
 import com.mg105.entities.GameState;
