+++ src/test/java/com/mg105/usecase/Inventory/InventoryInteractorTest.java
@@ -8,6 +8,7 @@ import com.mg105.entities.items.HealthPotion;
 import com.mg105.entities.items.UpgradeToken;
 import com.mg105.outputds.ItemDetails;
 import com.mg105.presenter_interfaces.InventoryPresenterInterface;
+import com.mg105.use_cases.inventory.InventoryInteractor;
 import com.mg105.utils.ItemConstants;
 import org.junit.jupiter.api.Test;
